Move resources to data module
